Extract generic config helpers
