+++ jq/src/main/java/se/code77/jq/config/simple/EventDispatcher.java
@@ -1,0 +1,22 @@
+package se.code77.jq.config.simple;
+
+import se.code77.jq.config.Config;
+
+public class EventDispatcher implements Config.Dispatcher {
+    private final EventThread mThread;
+
+    public EventDispatcher(EventThread t) {
+        mThread = t;
+    }
+
+    @Override
+    public void dispatch(Runnable r) {
+        dispatch(r, 0);
+    }
+
+    @Override
+    public void dispatch(Runnable r, long ms) {
+        mThread.post(r, ms);
+    }
+
+}
